requirements.txt
bs4==0.0.1
Pillow==7.1.1
requests==2.24.0
selenium==3.141.0
ulid-py==1.1.0
urllib3==1.25.8
